`decimal.DecimalContext` is final
